@@ -22,48 +22,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.9",
-    "@angular/cdk": "^6.4.7",
-    "@angular/common": "^6.1.0",
-    "@angular/compiler": "^6.1.0",
-    "@angular/core": "^6.1.0",
-    "@angular/forms": "^6.1.0",
-    "@angular/http": "^6.1.0",
-    "@angular/material": "^6.4.7",
-    "@angular/material-moment-adapter": "^6.4.7",
-    "@angular/platform-browser": "^6.1.0",
-    "@angular/platform-browser-dynamic": "^6.1.0",
-    "@angular/router": "^6.1.0",
+    "@angular/animations": "^7.0.0",
+    "@angular/cdk": "^7.0.0",
+    "@angular/common": "^7.0.0",
+    "@angular/compiler": "^7.0.0",
+    "@angular/core": "^7.0.0",
+    "@angular/forms": "^7.0.0",
+    "@angular/http": "^7.0.0",
+    "@angular/material": "^7.0.0",
+    "@angular/material-moment-adapter": "^7.0.0",
+    "@angular/platform-browser": "^7.0.0",
+    "@angular/platform-browser-dynamic": "^7.0.0",
+    "@angular/router": "^7.0.0",
     "angular-material-icons": "^0.7.1",
     "angular2-text-mask": "^9.0.0",
-    "angular4-material-table": "^0.6.8",
-    "core-js": "^2.5.4",
-    "husky": "^1.1.1",
+    "core-js": "^2.5.7",
+    "husky": "^1.1.2",
     "lint-staged": "^7.3.0",
     "moment": "^2.22.2",
-    "rxjs": "~6.2.0",
+    "rxjs": "~6.3.3",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.8.0",
-    "@angular/cli": "~6.2.3",
-    "@angular/compiler-cli": "^6.1.0",
-    "@angular/language-service": "^6.1.0",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.3.0",
-    "jasmine-core": "~2.99.1",
+    "@angular-devkit/build-angular": "~0.10.2",
+    "@angular/cli": "~7.0.2",
+    "@angular/compiler-cli": "^7.0.0",
+    "@angular/language-service": "^7.0.0",
+    "@types/jasmine": "~2.8.9",
+    "@types/jasminewd2": "~2.0.5",
+    "@types/node": "~10.12.0",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~3.2.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-coverage-istanbul-reporter": "~2.0.4",
     "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
+    "karma-jasmine-html-reporter": "^1.3.1",
     "prettier": "1.14.3",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
+    "protractor": "~5.4.1",
+    "ts-node": "~7.0.1",
     "tslint": "~5.11.0",
-    "typescript": "~2.9.2"
+    "typescript": "~3.1.3"
   }
 }
