@@ -1,13 +1,13 @@
 import { TestBed } from "@angular/core/testing";
 
-import { PersonDataProviderService } from "./person-data-provider.service";
+import { PersonInMemoryDataProviderService } from "./person-in-memory-data-provider.service";
 
 describe("PersonDataProviderService", () => {
   beforeEach(() => TestBed.configureTestingModule({}));
 
   it("should be created", () => {
-    const service: PersonDataProviderService = TestBed.get(
-      PersonDataProviderService
+    const service: PersonInMemoryDataProviderService = TestBed.get(
+      PersonInMemoryDataProviderService
     );
     expect(service).toBeTruthy();
   });
